feat: move contributions and logic in ai-terminal-summary-addons extension to ai-ide extension
