+++ examples/browser/package.json
@@ -45,7 +45,6 @@
     "@theia/ai-scanoss": "1.70.0",
     "@theia/ai-terminal": "1.70.0",
     "@theia/ai-terminal-assistant": "1.70.0",
-    "@theia/ai-terminal-summary-addons": "1.70.0",
     "@theia/ai-vercel-ai": "1.70.0",
     "@theia/api-provider-sample": "1.70.0",
     "@theia/api-samples": "1.70.0",
