+++ examples/browser/tsconfig.json
@@ -81,9 +81,6 @@
       "path": "../../packages/ai-terminal-assistant"
     },
     {
-      "path": "../../packages/ai-terminal-summary-addons"
-    },
-    {
       "path": "../../packages/ai-vercel-ai"
     },
     {
