+++ examples/electron/tsconfig.json
@@ -78,7 +78,7 @@
       "path": "../../packages/ai-terminal"
     },
     {
-      "path": "../../packages/ai-terminal-summary-addons"
+      "path": "../../packages/ai-terminal-assistant"
     },
     {
       "path": "../../packages/ai-vercel-ai"
