+++ packages/ai-ide/package.json
@@ -20,6 +20,7 @@
     "@theia/ai-core": "1.70.0",
     "@theia/ai-mcp": "1.70.0",
     "@theia/ai-terminal": "1.70.0",
+    "@theia/ai-terminal-assistant": "1.70.0",
     "@theia/core": "1.70.0",
     "@theia/debug": "1.70.0",
     "@theia/editor": "1.70.0",
