+++ packages/ai-terminal-assistant/src/browser/ai-terminal-assistant-command-service.ts
@@ -1,15 +1,15 @@
 import { Command, CommandRegistry } from '@theia/core';
 import { inject, injectable, postConstruct } from '@theia/core/shared/inversify';
 
-export interface AiTerminalSummaryCommandService {
+export interface AiTerminalAssistantCommandService {
     commands: Command[];
     executeCommand(commandId: string, ctx: any): void;
 }
 
-export const AiTerminalSummaryCommandService = Symbol('AiTerminalSummaryCommandService');
+export const AiTerminalAssistantCommandService = Symbol('AiTerminalAssistantCommandService');
 
 @injectable()
-export class AiTerminalSummaryCommandServiceImpl implements AiTerminalSummaryCommandService {
+export class AiTerminalAssistantCommandServiceImpl implements AiTerminalAssistantCommandService {
     @inject(CommandRegistry)
     protected readonly commandRegistry: CommandRegistry;
 
@@ -17,15 +17,15 @@ export class AiTerminalSummaryCommandServiceImpl implements AiTerminalSummaryCom
 
     @postConstruct()
     protected init(): void {
-        this.commands = this.getAiTerminalCommands();
+        this.commands = this.getAiTerminalAssistantCommands();
     }
 
     executeCommand(commandId: string, ctx: any): void {
         this.commandRegistry.executeCommand(commandId, ctx);
     }
 
-    protected getAiTerminalCommands(): Command[] {
+    protected getAiTerminalAssistantCommands(): Command[] {
         const commands = this.commandRegistry.commands;
-        return commands.filter(command => command.category === 'Ai Terminal');
+        return commands.filter(command => command.category === 'Ai Terminal Assistant');
     }
 }
