--- packages/ai-terminal-summary-addons/.eslintrc.js
@@ -1,10 +1,0 @@
-/** @type {import('eslint').Linter.Config} */
-module.exports = {
-    extends: [
-        '../../configs/build.eslintrc.json'
-    ],
-    parserOptions: {
-        tsconfigRootDir: __dirname,
-        project: 'tsconfig.json'
-    }
-};
