--- packages/ai-terminal-summary-addons/README.md
@@ -1,31 +1,0 @@
-<div align='center'>
-
-<br />
-
-<img src='https://raw.githubusercontent.com/eclipse-theia/theia/master/logo/theia.svg?sanitize=true' alt='theia-ext-logo' width='100px' />
-
-<h2>ECLIPSE THEIA - AI TERMINAL SUMMARY ADDONS EXTENSION</h2>
-
-<hr />
-
-</div>
-
-## Description
-
-The `@theia/ai-terminal-summary-addons` extension contributes commands to the terminal summary extension.\
-It enables the @theia/ai-chat extension to be automatically prompted with the terminal summary output.
-
-## Additional Information
-
-- [Theia - GitHub](https://github.com/eclipse-theia/theia)
-- [Theia - Website](https://theia-ide.org/)
-
-## License
-
-- [Eclipse Public License 2.0](http://www.eclipse.org/legal/epl-2.0/)
-- [一 (Secondary) GNU General Public License, version 2 with the GNU Classpath Exception](https://projects.eclipse.org/license/secondary-gpl-2.0-cp)
-
-## Trademark
-
-"Theia" is a trademark of the Eclipse Foundation
-<https://www.eclipse.org/theia>
