--- packages/ai-terminal-summary-addons/package.json
@@ -1,52 +1,0 @@
-{
-  "name": "@theia/ai-terminal-summary-addons",
-  "version": "1.70.0",
-  "description": "Theia - AI Terminal Summary Addons Extension",
-  "dependencies": {
-    "@theia/ai-chat": "1.70.0",
-    "@theia/ai-core": "1.70.0",
-    "@theia/ai-terminal-assistant": "1.70.0",
-    "@theia/core": "1.70.0",
-    "@theia/terminal": "1.70.0",
-    "@theia/debug": "1.70.0"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "theiaExtensions": [
-    {
-      "frontend": "lib/browser/ai-terminal-summary-addons-frontend-module"
-    }
-  ],
-  "keywords": [
-    "theia-extension"
-  ],
-  "license": "EPL-2.0 OR GPL-2.0-only WITH Classpath-exception-2.0",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/eclipse-theia/theia.git"
-  },
-  "bugs": {
-    "url": "https://github.com/eclipse-theia/theia/issues"
-  },
-  "homepage": "https://github.com/eclipse-theia/theia",
-  "files": [
-    "lib",
-    "src"
-  ],
-  "scripts": {
-    "build": "theiaext build",
-    "clean": "theiaext clean",
-    "compile": "theiaext compile",
-    "lint": "theiaext lint",
-    "test": "theiaext test",
-    "watch": "theiaext watch"
-  },
-  "devDependencies": {
-    "@theia/ext-scripts": "1.70.0"
-  },
-  "nyc": {
-    "extends": "../../configs/nyc.json"
-  },
-  "gitHead": "21358137e41342742707f660b8e222f940a27652"
-}
