--- packages/ai-terminal-summary-addons/src/browser/style/ai-terminal-summary-addons.css
@@ -1,8 +1,0 @@
-.theia-button {
-    height: 100%;
-    box-sizing: border-box;
-    font-size: 12px;
-    padding: 2px 4px;
-    min-width: 40px;
-    margin: 0;
-}
