--- packages/ai-terminal-summary-addons/tsconfig.json
@@ -1,31 +1,0 @@
-{
-  "extends": "../../configs/base.tsconfig",
-  "compilerOptions": {
-    "composite": true,
-    "rootDir": "src",
-    "outDir": "lib"
-  },
-  "include": [
-    "src"
-  ],
-  "references": [
-    {
-      "path": "../ai-chat"
-    },
-    {
-      "path": "../ai-core"
-    },
-    {
-      "path": "../ai-terminal-assistant"
-    },
-    {
-      "path": "../core"
-    },
-    {
-      "path": "../debug"
-    },
-    {
-      "path": "../terminal"
-    }
-  ]
-}
